+++ CSword.cpp
@@ -1,6 +1,6 @@
 #include "CSword.h"
 
-CSword::CSword(std::string p_nom, int p_damage,int p_bonus,int p_criticalStrike , int p_durability)
+CSword::CSword(std::string p_nom, int p_damage,int p_bonus,float p_criticalStrike , int p_durability)
 {
     m_nom = p_nom;
     m_damage = p_damage;
