+++ CWeapon.h
@@ -11,10 +11,10 @@ class CWeapon
     public:
         int m_damage;
         int m_bonus;
-        int m_criticalStrike;
+        float m_criticalStrike;
         std::string m_nom;
         CWeapon();
-        CWeapon(std::string p_nom, int p_damage,int p_bonus,int p_criticalStrike);
+        CWeapon(std::string p_nom, int p_damage,int p_bonus,float p_criticalStrike);
         virtual ~CWeapon();
         bool criticalStrike();
